### 2022-09-13 #4758 @danielskatz
👋 @linuxscout - the [paper.md in your repo](https://github.com/linuxscout/pyarabic/blob/master/paper.md) appears empty. We cannot proceed without a paper.

Please follow the [example paper](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography) and note that you can click on the error above to find out more about it. Please feel free to make changes to your .md and .bib files, then use the command `@editorialbot check references` to check the references, and the command `@editorialbot generate pdf` after making changes to the .md file or the .bib file to make a new PDF. editorialbot commands need to be the first entry in a new comment.
