### 2022-09-13 #4758 @danielskatz
👋 @linuxscout - note that your paper does not compile. Please follow the [example paper](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography) and note that you can click on the error above to find out more about it. It appears that you should change the affiliation in line 12 from `[1]` to just `1`.  In addition, you could work on the possibly missing DOIs that editorialbot suggests, but note that some may be incorrect. Also, if any of these are in your .bib file but not actually used in this paper, please remove them. Please feel free to make changes to your .bib file, then use the command `@editorialbot check references` to check again, and the command `@editorialbot generate pdf` after making changes to the .md file or when the references are right to make a new PDF. editorialbot commands need to be the first entry in a new comment.
