+++ paper.bib
@@ -16,7 +16,7 @@
   title={Semantic Word Clustering from Large Arabic Text},
   author={Abufayad, Tareq Issa},
   year={2018},
-  school={The Islamic University of Gaza}
+  institution={The Islamic University of Gaza}
 }
 @Article{abozinadah2016improved:2016,
   title		= {Improved microblog classification for detecting abusive
@@ -27,6 +27,7 @@
   year		= {2016},
   number	= {6},
   pages		= {17--28},
+  doi={10.5121/ijdkp.2016.6602},
   volume	= {6}
 }
 
@@ -40,6 +41,7 @@
   number={4},
   pages={544--559},
   year={2020},
+  doi={10.1177/0165551519849516},
   publisher={SAGE Publications Sage UK: London, England}
 }
 @inproceedings{kaibi2019comparative:2019,
@@ -48,24 +50,29 @@
   booktitle={2019 International Conference on Wireless Technologies, Embedded and Intelligent Systems (WITS)},
   pages={1--4},
   year={2019},
+  doi={10.1109/wits.2019.8723864},
   organization={IEEE}
 }
 
 
+
 @article{hamed2017building:2017,
-  title={Building a first language model for code-switch Arabic-English},
-  author={Hamed, Injy and Elmahdy, Mohamed and Abdennadher, Slim},
-  journal={Procedia Computer Science},
-  volume={117},
-  pages={208--216},
-  year={2017}
+title = {Building a First Language Model for Code-switch Arabic-English},
+journal = {Procedia Computer Science},
+volume = {117},
+pages = {208-216},
+year = {2017},
+note = {Arabic Computational Linguistics},
+issn = {1877-0509},
+doi = {https://doi.org/10.1016/j.procs.2017.10.111},
+url = {https://www.sciencedirect.com/science/article/pii/S1877050917321683},
+author = {Injy Hamed and Mohamed Elmahdy and Slim Abdennadher},
+keywords = {Automatic Speech Recognition, language model, code-mixing, code-switching, Arabic-English corpus, web corpus, web crawling},
[…]
