+++ paper.bib
@@ -64,7 +64,7 @@ pages = {208-216},
 year = {2017},
 note = {Arabic Computational Linguistics},
 issn = {1877-0509},
-doi = {https://doi.org/10.1016/j.procs.2017.10.111},
+doi = {10.1016/j.procs.2017.10.111},
 url = {https://www.sciencedirect.com/science/article/pii/S1877050917321683},
 author = {Injy Hamed and Mohamed Elmahdy and Slim Abdennadher},
 keywords = {Automatic Speech Recognition, language model, code-mixing, code-switching, Arabic-English corpus, web corpus, web crawling},
