+++ paper.md
@@ -9,7 +9,7 @@ tags:
 authors:
   - name: Taha Zerrouki
     email: t.zerrouki@univ-bouira.dz
-    affiliation: [1]
+    affiliation:1
     orcid: 0000-0001-9960-5533
     corresponding: true
 affiliations:
