+++ paper.md
@@ -9,12 +9,12 @@ tags:
 authors:
   - name: Taha Zerrouki
     email: t.zerrouki@univ-bouira.dz
-    affiliation: 1
+    affiliation: [1]
     orcid: 0000-0001-9960-5533
     corresponding: true
 affiliations:
-    index: 1
     name: Bouira University, Bouira, Algeria
+    index: 1
 date: 12 September 2022
 bibliography: paper.bib
 ---
