### 2022-10-27 #4886 @editorialbot
[REVIEW]: PyArabic: A Python package for Arabic text
**Submitting author:** @linuxscout (<a href="http://orcid.org/0000-0001-9960-5533">Taha Zerrouki</a>)
**Repository:** https://github.com/linuxscout/pyarabic
**Branch with paper.md** (empty if default branch): paper
**Version:** 0.6.16
