### 2023-01-06 #4886 @kikarimullah
@linuxscout

**State of the field**: Do the authors describe how this software compares to other commonly-used packages?

I think it would be worthwhile having a separate section in the paper where you address other software packages that are out there. There are many Arabic tokenizers, for example, and a lot of research has been done comparing them. It would be worth saying briefly how yours compares to them. See, for example, the many publications by Nizar Habash and his collaborators.
