### 2023-01-06 #4886 @kikarimullah
@linxscout:

**Quality of writing**: Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)?
Lines 13-14: "Before using machine learning algorithms, the library becomes even more important for text processing." Sentence is unclear.

Line 21: "Text tokenization into tokens" Redundant?
Line 32: "The library can be found at PyPi.org index." Make this a footnote after the first sentence of section Statement of Need.
Line 41: "arabic" Arabic (capitalise)
Lines 34-42 Please consistently capitalise after colon.
Lines 60-70 Please consistently capitalise in lists.
