+++ paper.md
@@ -28,17 +28,16 @@ PyArabic is a collection of modules that provide basic functionality for manipul
 It primarily provides preprocessing tools such as normalization, tokenization, diacritics removal, number conversion, transliteration, and so on.
 
 
-For years, researchers and developers have used the library to process and preprocess texts in their projects.
-Before using machine learning algorithms, the library becomes even more important for text processing. 
+For years, researchers and developers who worked on machine learning algorithms for natural language processing have used the library for Arabic text preprocessing and cleaning. The library becomes more important for machine learning.
 
 
 # Statement of need
 
-PyArabic is a Natural Language Processing Python package for Arabic text.
+PyArabic is a Natural Language Processing Python package for Arabic text\footnote{The library can be found at [PyPi.org index](https://pypi.org/project/PyArabic/)}.
 It is a simple library with basic functions for manipulating Arabic letters and text, such as detecting Arabic letters, Arabic letter groups and characteristics, removing diacritics, and so on.
 It contains the most basic and useful routines used by developers and researchers working with Arabic texts. Some key features are as follows:
 
--   Text tokenization into tokens.
+-   Text tokenization.
 -   Remove diacriticts (Harakat) from words (all, except Shadda, Tatweel, last haraka).
 -   Separate a word into letters and diacritics.
 -   Reduce diactritics of words.
@@ -50,19 +49,19 @@ It contains the most basic and useful routines used by developers and researcher
 -   Convert encoding and transliteration.
 
 
-The library can be found at [PyPi.org index](https://pypi.org/project/PyArabic/) \footnote{\url{https://pypi.org/project/PyArabic/}}.
+
 
 The PyArabic package includes five major submodules:
 
 - Araby: Basic tools and routines for manipulating Arabic text and letters, such as tokenization and diacritics removal, are provided.
-- Number: contains routines for dealing with numbers and numeric words; allows conversion of numbers to words and words to numbers; detects numeric phrases, and more.
-- Named provides simple tools for extracting named entities from text.
-- Trans: provides functions for converting between Arabic transliterations such as SAMPA, TIM Bukwalter, and Unicode.
-- Normalize: Utility functions that are used to prepare an arabic text for searching and indexing.
+- Number: Contains routines for dealing with numbers and numeric words; allows conversion of numbers to words and words to numbers; detects numeric phrases, and more.
+- Named: Provides simple tools for extracting named entities from text.
+- Trans: Provides functions for converting between Arabic transliterations such as SAMPA, TIM Bukwalter, and Unicode.
+- Normalize: Utility functions that are used to prepare an Arabic text for searching and indexing.
 
 More advanced projects use PyArabic, such as:
 
-- Adawat is an open framework for processing Arabic language that the author developed as part of his PhD research. In this work, we release a set of tools, the most important of which are:
+- Adawat is an open framework for processing Arabic language that the author developed as part of his PhD research. In PhD work, we release a set of tools, the most important of which are:
 
 - Mishkal, for restoring Arabic text diacritics [@mishkal].
 - Qalsadi is an Arabic morphology analyzer [@qalsadi].
[…]
