### 2023-01-06 #4886 @linuxscout
@kikarimullah
Thanks for you comments, I fix  them.

> @linxscout:
>
> **Quality of writing**: Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)? Lines 13-14: "Before using machine learning algorithms, the library becomes even more important for text processing." Sentence is unclear.

Fixed as: For years, researchers and developers who worked on machine learning algorithms for natural language processing have used the library for Arabic text preprocessing and cleaning. The library becomes more important for machine learning.

> Line 21: "Text tokenization into tokens" Redundant?

Fixed

>Line 32: "The library can be found at PyPi.org index." Make this a
footnote after the first sentence of section Statement of Need.

Fixed

>Line 41: "arabic" Arabic (capitalise)
Fixed

>Lines 34-42 Please consistently capitalise after colon.

Fixed
>Line 45: "In this work, we release a set of tools" Does 'this' refer to your PhD research or to this manuscript under review? Can you please disambiguate?

Fixed as: in  PhD word,

>  Lines 60-70 Please consistently capitalise in lists.

Fixed.
