+++ paper.bib
@@ -468,3 +468,79 @@ doi = {10.18653/v1/2021.eacl-main.204}
   journal = {GitHub repository},
   url = {https://github.com/linuxscout/qalsadi}
 }
+@misc{ar-php,
+  title={Ar-PHP, PHP library for website developers to process Arabic content},
+  author={Al-Shamaa, Khaled},
+  url={https://github.com/khaled-alshamaa/ar-php},
+  version = {6.3.1},
+  year={2022}
+}
+@inproceedings{obeid-etal-2020-camel,
+   title = "{CAM}e{L} Tools: An Open Source Python Toolkit for {A}rabic Natural Language Processing",
+   author = "Obeid, Ossama  and
+      Zalmout, Nasser  and
+      Khalifa, Salam  and
+      Taji, Dima  and
+      Oudah, Mai  and
+      Alhafni, Bashar  and
+      Inoue, Go  and
+      Eryani, Fadhl  and
+      Erdmann, Alexander  and
+      Habash, Nizar",
+   booktitle = "Proceedings of the 12th Language Resources and Evaluation Conference",
+   month = may,
+   year = "2020",
+   address = "Marseille, France",
+   publisher = "European Language Resources Association",
+   url = "https://www.aclweb.org/anthology/2020.lrec-1.868",
+   pages = "7022--7032",
+   abstract = "We present CAMeL Tools, a collection of open-source tools for Arabic natural language processing in Python. CAMeL Tools currently provides utilities for pre-processing, morphological modeling, Dialect Identification, Named Entity Recognition and Sentiment Analysis. In this paper, we describe the design of CAMeL Tools and the functionalities it provides.",
+   language = "English",
+   ISBN = "979-10-95546-34-4",
+}
+@misc{tkseem2020,
+  author = {Zaid Alyafeai and Maged Saeed},
+  title = {tkseem: A Tokenization Library for Arabic.},
+  year = {2020},
+  publisher = {GitHub},
+  journal = {GitHub repository},
+  howpublished = {\url{https://github.com/ARBML/tkseem}}
+}
+
+@misc{tnkeeh2020,
+  author = {Zaid Alyafeai and Maged Saeed},
+  title = {tkseem: A Preprocessing Library for Arabic.},
+  year = {2020},
+  publisher = {GitHub},
+  journal = {GitHub repository},
+  howpublished = {\url{https://github.com/ARBML/tnkeeh}}
+}
+
+@inproceedings{pasha2014madamira,
+  title={Madamira: A fast, comprehensive tool for morphological analysis and disambiguation of arabic},
+  author={Pasha, Arfath and Al-Badrashiny, Mohamed and Diab, Mona and El Kholy, Ahmed and Eskander, Ramy and Habash, Nizar and Pooleery, Manoj and Rambow, Owen and Roth, Ryan},
+  booktitle={Proceedings of the ninth international conference on language resources and evaluation (LREC'14)},
+  pages={1094--1101},
+  year={2014}
+}
+@inproceedings{abdelali2016farasa,
+  title={Farasa: A fast and furious segmenter for arabic},
+  author={Abdelali, Ahmed and Darwish, Kareem and Durrani, Nadir and Mubarak, Hamdy},
+  booktitle={Proceedings of the 2016 conference of the North American chapter of the association for computational linguistics: Demonstrations},
+  pages={11--16},
+  year={2016}
+}
+@article{loper2002nltk,
+  title={Nltk: The natural language toolkit},
+  author={Loper, Edward and Bird, Steven},
+  journal={arXiv preprint cs/0205028},
+  year={2002}
+}
+@book{vasiliev2020natural,
[…]
