### 2023-01-06 #4886 @linuxscout
> @linuxscout
>
> **State of the field**: Do the authors describe how this software compares to other commonly-used packages?
>
> I think it would be worthwhile having a separate section in the paper where you address other software packages that are out there. There are many Arabic tokenizers, for example, and a lot of research has been done comparing them. It would be worth saying briefly how yours compares to them. See, for example, the many publications by Nizar Habash and his collaborators. In the documentation, you also credit Khaled Alshamaa PHP Project. Perhaps you could say a bit more about how PyArabic relates to this project in the paper?

Hi @kikarimullah,
I added this section into paper:

PyArabic was inspired by Ar-PHP[@ar-php], an Arabic library for the PHP programming language that provides basic routines for web developers. Then the two libraries grow together through collaborations, and they are inspired mutually by each other. Ar-PHP provides basic routines for PHP and MySQL databases and attempts to solve web development issues such as arabic glyph rendering; however, the Ar-PHP library also includes advanced modules such as sentiment analysis, muslim prayer times, and auto-summarize [@ar-php].

There are many dedicated frameworks for Arabic natural language processing, like MADAMIRA(Java) [@pasha2014madamira], FARASA(Java)[@abdelali2016farasa], CAMeL(Python) [@obeid-etal-2020-camel]. Many multilingual frameworks, however, such as NLTK (Python) [@loper2002nltk], Spacy (Python) [@vasiliev2020natural], and CLTK (Python) [@johnson2014:2014], only partially support Arabic.

In PyArabic, we focused on basic routines and build our library to be native and independent enough to be embedded in complex projects. This library was used in many projects and adopted by frameworks like CLTK[@johnson2014:2014], and has been inspired to build more specific libraries like TaKseem (a tokenization library for Arabic) [@tkseem2020] and Tankeeh (Arabic cleaning, normalization, and segmentation library) [@tnkeeh2020].
