### 2023-01-06 #4886 @kikarimullah
@linuxscout

RE: DOCUMENTATION
**Example usage**: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).

There are examples showing the input and output of the different functions. The examples are a bit contrived, however. Since this is an Arabic NLTK, I wonder whether using larger text samples from naturally occurring texts (circa 10^2 or 10^3 words?), perhaps from Twitter, perhaps from Arabic news sites, perhaps from classical Arabic texts sourced from al-maktabah.org or from Maktaba Shamela?

This may be valuable in for those wanting to use the tokenization functions to generate frequency lists, for example.

This is a suggestion for consideration.
