### 2023-01-06 #4886 @kikarimullah
@linuxscout

RE: DOCUMENTATION

Currently the documentation is inconsistently translated into English. Sometimes there is a brief translation, other times English translation is almost entirely absent. I'd recommend ensuring that English translations are consistently provided in some way (Khaled Alshamaa has separate Arabic and English sites for his repos). Consistent English translation will *definitely* widen the scope of potential users of the software.

This is a suggestion for your consideration.
