### 2023-01-06 #4886 @kikarimullah
RE: DOCUMENTATION

There is a problem with the example for the function unshaping_text:
Unshape a text: https://pyarabic.readthedocs.io/ar/latest/features.html#id7

Part of the first bayt of the poem is missing.

This is a suggestion for your consideration.
