### 2023-01-09 #4886 @linuxscout
> @linuxscout
>
> RE: DOCUMENTATION **Example usage**: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).
>
> There are examples showing the input and output of the different functions. The examples are a bit contrived, however. Since this is an Arabic NLP tookkit, I wonder whether using larger text samples from naturally occurring texts (circa 10^2 or 10^3 words?), perhaps from Twitter, perhaps from Arabic news sites, perhaps from classical Arabic texts sourced from al-maktabah.org or from Maktaba Shamela?

Examples in documentation are used as illustrations to show functionalities. The library is used to solve real world problems within other packages, like ours (Mishkal, Qutrub, and Qalsadi), which give us feedback to improve the pyarabic.In project Mishkal, for example, we overuse the strip_tashkeel function; profiling Mishkal programs led us to select the fastest algorithm for removing diacritics.

>
> This may be valuable in for those wanting to use the tokenization functions to generate frequency lists, for example.
>

The tokenize fucntion has advanced options, It can be used with other  conditions  like (restrict Arabic, keep or remove numbers, exclude stop words ...etc).

For example , to remove tashkeel and filter out non-Arabic words, we use this code

```python
>>> from pyarabic.araby import tokenize, is_arabicrange, strip_tashkeel
>>> text = u"ِاسمٌ الكلبِ في اللغةِ الإنجليزية Dog واسمُ الحمارِ Donky"
>>> tokenize(text, conditions=is_arabicrange, morphs=strip_tashkeel)
        ['اسم', 'الكلب', 'في', 'اللغة', 'الإنجليزية', 'واسم', 'الحمار']

```

>  functions to generate frequency lists,

I can be down with Counter lib, after tokenization.

> This is a suggestion for consideration.
