### 2023-01-09 0a31e83
explain some routines use
