### 2023-01-09 #4886 @linuxscout
Hi @kikarimullah,
> RE: SOFTWARE PAPER
>
> The paper chiefly consists in a series of lists. As a way of strengthening the analytical aspect of the paper, I think you might add a few more lines of prose when you speak about (1) advanced projects on lines 43 - 56 and (2) lines 60 - 70 where you mention how the software is currently being used.
>
> For (1) you could add a paragraph that gives an overall characterisation of how certain functions work in more advanced NLP software. Perhaps you could briefly discuss an example, say, of how Qutrub or Qalsadi depend on PyArabic libraries.

done

For (2) you could discuss why, for example, PyArabic has been used so extensively in research on sentiment analysis.

I mention that PyArabic was created to aid researchers and developers in natural language processing tasks, particularly text preprocessing (tokenization, cleaning, normalization, strip diacritics). It has already appeared in several scientific publications. It is mentioned in:
