### 2023-04-04 #4886 @amitkumarj441
> 👋 @amitkumarj441 and @VenkteshV could I have an update please?

I am working on the review. I realised that the code licensing is GPLv3 in actual, though the link to license is broken on both GitHub (in both branches; paper/master)  and PyPi. I would suggest fixing these links on both GitHub repo and PyPi meta.
