+++ README.md
@@ -16,7 +16,7 @@ Features |   value
 ---------|---------------------------------------------------------------------------------
 Authors  | Taha Zerrouki: http://tahadz.com,  taha dot zerrouki at gmail dot com
 Release  | 0.6.12 
-License  |[GPL](https://github.com/linuxscout/pyarabic/master/LICENSE)
+License  |[GPL](https://github.com/linuxscout/pyarabic/blob/master/LICENSE)
 Tracker  |[linuxscout/pyarabic/Issues](https://github.com/linuxscout/pyarabic/issues)
 Website  |[https://pypi.python.org/pypi/pyarabic](https://pypi.python.org/pypi/pyarabic)
 Doc  |[package Documentaion](https://pyarabic.readthedocs.io/)
