### 2023-04-14 #4886 @amitkumarj441
I have tested this library functionalities and almost every [features](https://github.com/linuxscout/pyarabic/blob/master/doc/features.md) works with the python tests without a hitch. Though, there multiple folders for docs (`doc` and `docs`), keeping both would be okay as long as the main `MANIFEST` file contains `include docs/*` & `include tests/*`. Also, for mainstream documentation page: https://pyarabic.readthedocs.io/ar/latest/ for which the configuration file (`.readthedocs.yml`) is missing.

I am happy to recommend the paper for publication at this point provided the above suggestions should be considered and hope that @linuxscout can provide their feedback soon.
