+++ paper.bib
@@ -525,6 +525,7 @@ doi = {10.18653/v1/2021.eacl-main.204}
 }
 @inproceedings{abdelali2016farasa,
   title={Farasa: A fast and furious segmenter for arabic},
+  doi={10.18653/v1/N16-3003},
   author={Abdelali, Ahmed and Darwish, Kareem and Durrani, Nadir and Mubarak, Hamdy},
   booktitle={Proceedings of the 2016 conference of the North American chapter of the association for computational linguistics: Demonstrations},
   pages={11--16},
