### 2023-04-15 #4886 @linuxscout
Hi @amitkumarj441,
Thank you a lot for your valuable comments,
I updated my repo and fix doi missing.
Thanks again.
 > Though, there multiple folders for docs (`doc` and `docs`), keeping both would be okay as long as the main `MANIFEST` file contains `include docs/*` & `include tests/*`. Also, for mainstream documentation page: https://pyarabic.readthedocs.io/ar/latest/ for which the configuration file (`.readthedocs.yml`) is missing, including missing DOIs needs to be fixed.
>
> I am happy to recommend the paper for publication at this point provided the above suggestions should be considered and hope that @linuxscout can provide their feedback soon.
