### 2023-04-16 #4886 @linuxscout
Hi @ajstewartlang,
Thank you for your recommendations,
I would to know if upload the package into Zenodo as publication or as a software.
Thanks
