### 2023-04-16 #4886 @ajstewartlang
> Hi @ajstewartlang, Thank you for your recommendations, I would to know if upload the package into Zenodo as publication or as a software. Thanks

Hi @linuxscout - it's the tagged release of the reviewed version of the software that you want to archive on Zenodo.
