### 2023-04-20 #4886 @arfon
@linuxscout – could you please merge this PR? https://github.com/linuxscout/pyarabic/pull/71
