### 2023-04-20 c3ceab6
Update Citation
Add citation into ```Zerrouki, T., (2023). PyArabic: A Python package for Arabic text. Journal of Open Source Software, 8(84), 4886, https://doi.org/10.21105/joss.04886```
