### 2023-04-20 #4886 @linuxscout
Dear Journal of Open Source Software Editorial Team,

I am writing to express my sincere gratitude for accepting my paper, "PyArabic: A Python package for Arabic text," for publication in your esteemed journal. It is an honor to have my work featured in the Journal of Open Source Software, and I am thrilled at the opportunity to share my research with a wider audience.

I would also like to extend my heartfelt thanks to the reviewers who provided invaluable feedback and constructive criticism during the peer-review process. Their insightful comments have helped to strengthen the quality of my work, and I am grateful for their time and effort.

I believe that this publication will contribute to the advancement of the field of computational linguistics, particularly in the area of Arabic text analysis. With the availability of PyArabic, researchers and developers will have access to a powerful and user-friendly Python package that facilitates Arabic text processing and analysis.

Once again, I want to express my appreciation for the acceptance of my paper in the Journal of Open Source Software, and I look forward to continued collaboration with the journal in the future.

Sincerely,
